Add Binary event support for 'xml', 'test' and 'json' mapping payload.
