+++ component/src/test/java/org/wso2/extension/siddhi/io/kafka/sink/KafkaSinkTestCase.java
@@ -36,6 +36,9 @@ import java.rmi.RemoteException;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Class implementing the Test cases for Kafka Sink.
+ */
 public class KafkaSinkTestCase {
     static final Logger LOG = Logger.getLogger(KafkaSinkTestCase.class);
     private volatile int count;
